+++ src/go/cmdinit/init_command.go
@@ -1,4 +1,4 @@
-package cmd
+package cmdinit
 
 import (
 	"fmt"
